Added skip, getRemaining methods to Data IO classes. Added read only data buffers and positioned data super interface.
